favorites implementation

+++ MarsRoverPhotos/Model/CoreData/CDPhotos+CoreDataProperties.swift
@@ -16,10 +16,10 @@ extension CDPhotos {
         return NSFetchRequest<CDPhotos>(entityName: "CDPhotos")
     }
 
-    @NSManaged public var id: Int16
+    @NSManaged public var id: Int32
     @NSManaged public var roverName: String?
     @NSManaged public var addFavoritesDate: Date?
-    @NSManaged public var sol: Int16
+    @NSManaged public var sol: Int32
     @NSManaged public var earthDate: Date?
     @NSManaged public var cameraName: String?
